add config for solar river 4400
Seems like the samilpower config had mixed up
the MODE and ETODAY hex values for SR4400.

I dont know if it the same on the other Solar River
so I split this chaneg to a new file.

If someone can confim its the same problem on the other one
i can merge them again.

Signed-off-by: Robert Marklund <robbelibobban@gmail.com>